+++ lib/teiserver/contexts/connections.ex
@@ -35,7 +35,7 @@ defmodule Teiserver.Connections do
   @spec connect_user(Teiserver.user_id()) :: Client.t()
   defdelegate connect_user(user_id), to: ClientLib
 
-  @doc section: :client
+  @doc false
   @spec client_exists?(Teiserver.user_id()) :: pid() | boolean
   defdelegate client_exists?(user_id), to: ClientLib