+++ lib/teiserver/game/servers/lobby_id_server.ex
@@ -49,7 +49,7 @@ defmodule Teiserver.LobbyIdServer do
     end
   end
 
-  @spec start_link(List.t()) :: :ignore | {:error, any} | {:ok, pid}
+  @spec start_link(list) :: :ignore | {:error, any} | {:ok, pid}
   def start_link(_) do
     GenServer.start_link(__MODULE__, [], [])
   end
@@ -62,7 +62,7 @@ defmodule Teiserver.LobbyIdServer do
     {:noreply, %{state | next_id: next_id}}
   end
 
-  @spec init(Map.t()) :: {:ok, Map.t()}
+  @spec init(map()) :: {:ok, map()}
   def init(_opts) do
     Registry.register(
       Teiserver.ServerRegistry,