+++ test/game/lobby_server_test.exs
@@ -2,8 +2,8 @@ defmodule Teiserver.Game.LobbyServerTest do
   @moduledoc false
   use Teiserver.Case, async: true
 
-  alias Teiserver.Game
-  alias Teiserver.{AccountFixtures, ConnectionFixtures}
+  alias Teiserver.{Game, Connections}
+  alias Teiserver.ConnectionFixtures
 
   describe "Lobby server" do
     test "server lifecycle" do
@@ -17,10 +17,38 @@ defmodule Teiserver.Game.LobbyServerTest do
       assert lobby.host_id == user.id
       assert lobby.name == "Lobby #{r}}"
 
+      topic = Game.lobby_topic(lobby.id)
+      listener = TestConn.new(topic)
+
       # User disconnects
       TestConn.stop(conn)
 
-      flunk "In what way do we update the lobby to show the host is disconnected?"
+      [m] = TestConn.get(listener)
+
+      assert match?(
+               %{
+                 topic: ^topic,
+                 update_id: 1,
+                 event: :lobby_updated,
+                 lobby: %{
+                   host_data: nil
+                 }
+               },
+               m
+             )
+
+      # Now destroy the client process
+      Connections.stop_client_server(user.id)
+      # Sleep to let the process die
+      :timer.sleep(100)
+
+      [m] = TestConn.get(listener)
+
+      assert m == %{
+               event: :lobby_closed,
+               lobby_id: lobby_id,
+               topic: topic
+             }
     end
   end
 end